requirements.txt
torchvision
torch



# git+https://github.com/andrewangbl/myclip.git
git+https://github.com/openai/CLIP.git

pillow
numpy==1.24.3
fastapi
pydantic
python-dotenv
firebase-admin
uvicorn

ftfy
packaging
regex
tqdm
